refactor: unify sarama and env configs
